fix: split chart axis helpers to satisfy ESLint max-lines budget
Move cartesian axis formatting utilities out of WidgetChart.tsx so the
file stays under the 500-line lint threshold.
